create button to change font size
