--- src/routes/about/+page.svelte
@@ -1,14 +1,0 @@
-
-<h1>Settings</h1>
-
-
-<main class="main-page">
-
-
-
-
-</main>
-
-<footer>
- <p>&copy Andrey Marins</p>
-</footer>
