--- static/css/reader.css
@@ -1,13 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@100..900&display=swap');
-
-/*Definições de estilos*/
-
-:root{
-    --font-family: font-family: "Noto Sans JP", san-serif;
-    --text-size: 30px;
-}
-
-html{
-    font-size: var(--text-size);
-    font-family: var(--font-family);
-}
